Minimal Usage of the synchronized block For B+ Tree loading
Below listed are the points for which code modification is done.

1. Minimal usage synchronization  while table B-tree loading. Each query will list the folders which it has to load for fetching the result,
if the folders are already loaded the query will not enter the synchronized block which currently impacts the query performance.
2. After table loading, each query will form a query scope object which will define on which segments/load folders a query has to work on
which is helpful in concurrent scenarios.
3. Each segment cache object will maintain the modification time of segment which will help during retention scenario. Currently
after retention operation, the status of segments modified is marked for update becuase of which those segments were loaded each time the query
was executed. Now the modification time will be matched and the updated segment will be loaded only once.
4. Read write lock implementation has been done to allow concurrrent access of segment cache object list.
5. Added UT + modified some comments