Problem: Data loading fails if parsing a double value returns infinity
Analysis: During data load, if a value specified is too big for a double DataType column then while parsing that value as double result is returned as "Infinity". Due to this while we calculate min and max value for measures in carbon data writer step it throws an exception.

Fix: If result is Infinity or NAN for double value parsing then make the value as null and add it to bad records.

Impact area: Data load which contains non parseable values for a datatype.