Problem: Dictionary cache not getting cleared after task completion in dictionary decoder
Analysis: Currently LRU cache eviction policy is based on dictionary access count. For cache to remove a entry its access count must be 0. In dictionary decoder after conversion of surrogate key to actual value the access count for dictionary columns in query is not getting decremented due to which it will never be cleared from memory when LRU cache size is configured.

Fix: Add a task completion listener which will take care of clearing the dictionary in case of both success and failure

Impact area: LRU cache eviction policy which can lead to query and data load failure