Problem: Block distribution is wrong in case of dynamic allocation=true
Analysis: In case when dynamic allocation is true and configured max executors are more than the initial executors then carbon is not able to request the max number of executors configured. Due to this resources are getting under utilized and case when number of blocks increases, the distribution of blocks is limited to the number of nodes and the number of tasks launched are less. This leads to under utilization of resources and hence impacts the query and load performance.

Fix: Request for starting the maximum number of configured executors in case dynamic allocation is true.

Impact area: Query and data load flow performance due to under utilization of resources.