[CARBONDATA-400] * Problem: When the number of characters in a column exceeds 100000 characters whole string appears in beeline with exception.
Analysis: In univocity csv parser settings , the maximum number of characters per column is 100000 and when it exceeds that limit, TextparsingException is thrown with
the complete string as error in beeline during data load.

Fix: Now a proper error message is displayed in beeline and complete error messages and parser settings details will be present in logs.
Impact area: Data loading with more than 100000 characters in a single column.