Problem: Issue with database name case sensitivity
Analysis: When database name is provided in any DDL/DML command, the database name is interpreted and used in the same case as provided by the user. This leads to different behavior in windows and unix systems as windows is case sensitive and linux systems are case insensitive.
Consider a case for create database. Lets say database name is "Carbon". While executing database name is provided as Carbon but while deleting or using or creating table the case is changed to "CARbOn". In these cases system will not behave correctly and if HDFS UI is checked the database Carbon will still exist even after dropping database as the case for database name was different in the 2 commands execution.

Fix: Always convert the database name to lower case for any DDL/DML operation while parsing so that the behavior remains consistent.