Problem: Memory leak in case of compaction when unsafe is true
Analysis: In case of compaction, queryExecutor object is formed for multiple blocks but the objects are not retained and finish method is called only on the last instance created for query executor. Due to this the memory allocated to precious objects is not released which can lead to out of memory issue.

Fix: Add all the object to a list and call finish method on all the objects so that memory gets released