Problem: Dictionary generation failure if there is failure in closing output steam in HDFS
Analysis: If there is any failure while closing the output stream of dictionary file in HDFS then on next data load, update or insert into operation dictionary generation fails. This is because we open the dictionary file in append mode and when we try to get the output stream for that file HDFS throws an exception that Lease is already acquired by some other client.

Fix: Recover the lease through carbondata code if exception is for lease failure