[CARBONDATA-1429] Add a value based compression for decimal data type when decimal is stored as Int or Long
Added a value based compression for decimal data type when decimal is stored as Int or Long

When decimal precision is <= 9, decimal values are stored in 4 bytes but they are not compressed further based on min and max values as compared with other primitive data type compression. Therefore now based on min and max value decimal data falling in Integer range will be further compressed as byte or short.
When decimal precision is <= 18, decimal values are stored in 8 bytes but they are not compressed further based on min and max values as compared with other primitive data type compression. Therefore now based on min and max value decimal data falling in Long range will be further compressed as byte, short or int.
Advantage: This will reduce the storage space thereby decreasing the IO time while decompressing the data.

This closes #1297