Fixed issue of HDFS file system throwing exception while deleting a non-existent folder. this happens because on multiple run of insert overwrite operation, same folder is getting added to the stale folder list even though it has been deleted during first time insert overwrite operation run and when the same folder is tried to be deleted again HDFS file system throws an exception.
Insert or Load operation should throw exception if:
1. If insert overwrite is in progress and any other load or insert operation is triggered
2. If load or insert into operation is in progress and insert overwrite operation is triggered

Fixed issue for table status not getting updated after failure of insert/load job. In case of insert overwrite job status in table status file is "Overwrite in Progress". In this case no other load for same table can run in parallel and if this status is not changes on success of failure jobs, the system will still assume that insert overwrite is in progress and fail all next insert/load jobs

Fixed fortify issue for the category "Identical code for different branches"