[CARBONDATA-1509] Fixed bug for maintaining compatibility of decimal type with older releases of Carbondata
In old Carbondata releases, precision and scale is not stored for decimal data type and both values are initialized to -1. In TableSpec.ColumnSpec default values for precision and scale are initialized to 0 because of which exception is thrown while reading the old store with decimal column. Both precision and scale should be initialized to -1.

This closes #1378