[CARBONDATA-1844] Add tablePath support when creating table
User should be able to specify table path when creating table. This PR supports it.
The tablePath of the table is determined as following steps:

Use the table property 'table_path' specified by user
Use database 's locationUri if it is specified by user in create database command, and concatenate with table name as table path
Use carbon store path and database name and table name to construct a table path

This closes #1603