[CARBONDATA-1887] block pruning not happening is carbon for ShortType and SmallIntType columns
Block pruning not working for filter on Short and smallint type column.In PushedFilters it goes in CastExpr as below. PushedFilters: [IsNotNull(dob), CastExpr((cast(dob#230 as Integer) = 12))]. This is why it is getting pushed to spark.CastExpressionOptimization should handle the removing the cast from the expression for smallint and short type columns.

This closes #1647