[CARBONDATA-1862][PARTITION] Support compaction for partition table
It supports compaction on partition table.
There is a change in compaction during the block identification and grouping. As all blocks which are related same partition always needs to group to same set for compaction.So compactor needs to get the partition information from partition map file during compaction of partition table

This closes #1675