[CARBONDATA-1946] Exception thrown after alter data type change operation on dictionary exclude integer type column
Problem: After restructure change data type operation (INT to BIGINT) on dictionary exclude INT type column if select query is triggered then exception is thrown.

Analysis: This is happening because while retrieving the data the vector is created for BIGINT type (size 8 bytes) which but the actual length of each data is 4 bytes and there is length check while reading the data which is failing.

Solution: Added a new restructuredType variable in vector and assigned the block dimension data type to it.

This closes #1732