[CARBONDATA-1967][PARTITION] Fix autocompaction and auto merge index in partition tables
Auto compaction is not working in case of the partition table and merge index files are merging always even though it is configured as false.

Solution:
Auto compaction code is added after finishing of partition loading. And also merge index configuration is checked before going for index merging.

This closes #1748