[CARBONDATA-2020][Old Store Support] Add filter support for old store reading to improve query performance
Problem
For old stores blocklet level min/max comparison was not happening in the executor side due to which all the blocklets were getting scanned. This increased the IO and scanning time in the executor.

Solution
Modified code to retrieve the min/max value from blocklet node and use it for comparsion while scanning for valid blocklets.

This closes #1818