[CARBONDATA-2061] Check for only valid IN_PROGRESS segments
Problem:
During operations like drop, delete segment, compaction, IUD there is a check for the IN_PROGRESS segments of a table. This check is simply checking the tblstatus file for IN_PROGRESS segments.

Solution:
The check should validate the IN_PROGRESS segments and decide on Valid and Invalid IN_PROGRESS segments.

This closes #1844