[CI] Enabled pre aggregate tests and fixed insert query CI random query failure
Problem:
There is an issue of random null pointer exception from TaskContext while doing insert queries in CI that's why pre-aggregate tests are commented

Solution :
From my analysis, It is because of lack of synchronization in Sparks TaskContext and same is rectified in Spark 2.2.
To fix this issue added a mock to TaskContext and ignored the exceptions in tests.

This closes #1854