[CARBONDATA-2120]Fixed is null filter issue
Problem: Is null filter is failing for numeric data type(No dictionary column).

Root cause: Min max calculation is wrong when no dictionary column is not the first column.

As it is not the first column null value can come in between and min max for null value is getting updated only when first row is null

Solution: Update the min max in all the case when value is null or not null for all type

This closes #1912