[CARBONDATA-2183] Fix compaction when segment is delete during compaction and remove unnecessary parameters in functions
Problem:
when compaction is started and job is running, and parallelly the segment involved in the compaction is deleted using DeleteSegmentByID, then
compaction is success.

Solution:
when compaction is started and job is running, and parallelly the segment involved in the compaction is deleted using DeleteSegmentByID, then
compaction should be aborted and failed. and proper error message should thrown to user. THis PR also removes the unnecessary parameters in functions.

This closes #1979