[CARBONDATA-2199] Fixed Dimension column after restructure getting wrong block datatype
Problem: Changing datatype of measure having sort_columns calls for restructure and after having restructure it changes the datatype to actual datatype for which accessing the data with changed datatype gives exception of incorrect length.

Solution: Store the datatype in DimensionInfo while restructuring and access the same datatype to get the block data type.

This closes #1993