Fix exhausted memory problem during unsafe data loading
If the size of unsafe row page equals to that of working memory, the
last page will exhaust the working memory and carbondata will suffer 'not
enough memory' problem when we convert data to columnar format.

All unsafe pages should be spilled to disk or moved to unsafe sort memory
instead of be kept in unsafe working memory.