Merge and spill in-memory pages if memory is not enough
Currently in carbondata, pages will be added to memory. If memory is not
enough, newly incoming pages will be spilled to disk directly. This
implementation will merge&spill the in-memory pages and make room for
the newly incoming pages.
As a result, carbondata will spill less than before and spill bigger files
instead of smaller files.