[CARBONDATA-2275]Query Failed for 0 byte deletedelta file
When delete is failed on write step because of any exception from hdfs. Currently 0 bye deletedelta file is created and not getting cleaned up .
So when any Select Query is triggered , Select Query is failed

Root Cause:-
deletedelta of 0 size was considered for reader and while getting the hostname it has thrown IndexOutofBound.
Even Writer is failed while flush /close (exception from hdfs) but Exception is not thrown back to caller which is making delete operation is success.

Solution :-
While reading ignore 0 size deletedelta file When Writer is failed ,Throw Failure Exception to caller and show error on query client .
deletedelta of 0 byte will be handled in cleanup automatically in next delete/update .

This closes #2121