[CARBONDATA-2278] Save the datamaps to system folder of warehouse.
Make the datamap schema independent of main table schema. And store the schema under _system folder location. This location is configurable by using carbon property carbon.system.folder.location , by default, it stores under the store location.
Created datamap schema in JSON format for better readability. And has the interfaces to store it in database but not given any implementation to it in this PR.
Made on table <tablename> for datamap DDL as optional , so now user can create/drop or show datamaps without on table option.

This closes #2100