[CARBONDATA-2291] Added datamap status and refresh command to sync data manually to datamaps
In order maintain the data consistency we require to enable or disable datamaps when data is not synchronized between fact and datamap. So added the new status called datamapstatus file under the system folder. whenever the data is out of sync between parent tables and datamap it just updates the status as disabled in datamapstatus file. After user manually refreshes the datamap it will update the status to enable.
Only the enabled datamaps are considered during query.

This closes #2106